requirements.txt
redis
python-dotenv==0.11.0
python-telegram-bot==12.4.2
vk-api==11.7.0
